Moved logic to FollowedUsersSection and FollowButton Improved responsiveness
